Avoid calling `read-event'.
It's causing problems when it is called from a buffer switch hook which
is itself called in the middle of processing events. I also think we
actually dropped those events that we did read.

In my experiments, I've been seeing errors like the following happen a
lot with EXWM when switching back and forth between frames with the
mouse pointer:

    Debugger entered--Lisp error: (wrong-type-argument listp #<frame *GNU Emacs* 0x4718c00>)
      handle-select-window((switch-frame #<frame *GNU Emacs* 0x4718c00>))
      funcall-interactively(handle-select-window (switch-frame #<frame *GNU Emacs* 0x4718c00>))
      call-interactively(handle-select-window nil nil)
      command-execute(handle-select-window)

What's happening, as far as I can work out, is that we're in the middle
of reading an event from the keyboard when a buffer switch happens, and
the buffer switch hook calls exwm code which calls
`exwm--make-emacs-idle-for`, which in turn calls `read-event` and causes
Emacs to overwrite some of the global event data.